refactor: split runner.rs into focused submodules
Extract three modules from runner.rs (1297 LOC) to improve
maintainability and testability:

- listeners.rs: ExecutionListener trait + NullListener + StderrListener (54 LOC)
- json_parser.rs: parse_json_output() with 3 extraction strategies + tests (147 LOC)
- audit.rs: open_audit_log() + write_audit_entry() as free functions (56 LOC)

mod.rs remains the coordinator (~1142 LOC) with RecipeRunner,
run_steps(), execute_step(), dispatch_step(), parallel execution,
and hooks.

Public API preserved via re-exports — no breaking changes.
All 96 tests pass, clippy clean, rustfmt clean.
